+++ android/src/main/kotlin/vn/loi/learning/android/LearningEngineAndroidApplication.kt
@@ -4,5 +4,16 @@ import android.app.Application
 import vn.loi.learning.android.platform.AndroidApplicationGraph
 
 class LearningEngineAndroidApplication : Application() {
-    val graph: AndroidApplicationGraph by lazy { AndroidApplicationGraph.create(this) }
+    private val graphOwner by lazy {
+        SingleInstanceOwner { AndroidApplicationGraph.create(this) }
+    }
+
+    val graph: AndroidApplicationGraph
+        get() = graphOwner.value
+}
+
+internal class SingleInstanceOwner<T>(
+    create: () -> T
+) {
+    val value: T by lazy(LazyThreadSafetyMode.SYNCHRONIZED, create)
 }